# PluginBuilder/wwwroot/swagger/v1/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BTCPay Server Plugin Builder API",
    "description": "REST API for the BTCPay Server Plugin Builder. Provides endpoints for querying published plugin versions, downloading plugin binaries, checking for updates, and triggering new plugin builds.\n\nThe live server is available at [plugin-builder.btcpayserver.org](https://plugin-builder.btcpayserver.org/).",
    "version": "1.0.0",
    "contact": {
      "name": "BTCPay Server",
      "url": "https://github.com/btcpayserver/btcpayserver-plugin-builder"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://plugin-builder.btcpayserver.org/api/v1",
      "description": "Production"
    },
    {
      "url": "/api/v1",
      "description": "Current instance"
    }
  ],
...
  "paths": {
    "/admin/me": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Inspect current admin account",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/status": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Inspect server counts and feature flags",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "List users",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/users/{userId}": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Inspect a user",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/plugins": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "List all plugins",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/plugins/{pluginSlug}": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Inspect a plugin",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "pluginSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/plugins/{pluginSlug}/builds": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "List builds newest first",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "pluginSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/plugins/{pluginSlug}/builds/{buildId}": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Inspect build metadata and manifest",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "pluginSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "buildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/plugins/{pluginSlug}/builds/{buildId}/logs": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Read build log rows newest first",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "pluginSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "buildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/plugins/{pluginSlug}/builds/{buildId}/logs/{logId}": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "Read a log row in bounded chunks",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "pluginSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "buildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "logId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 2147467263,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
    "/admin/listing-requests": {
      "get": {
        "tags": [
          "Admin Agent"
        ],
        "summary": "List listing requests",
        "description": "Requires a current ServerAdmin account. See docs/admin-agent-access.md for schemas, cursor semantics and review/audit behavior.",
        "security": [
          {
            "Basic": []
          },
          {
            "AdminToken": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "all"
              ],
              "default": "pending"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Authentication failed."
          },
          "403": {
            "description": "Admin role required."
          },
          "404": {
            "description": "Resource not found."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Revocable admin access token issued in Account Settings. Admin API only; token management requires Basic auth."
      },
      "Basic": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic Authentication using your Plugin Builder login email and password. Format: `Authorization: Basic {base64(email:password)}`. Requires HTTPS in production."
      }
    },
...
  }
}